+++ js/components.js
@@ -9,6 +9,13 @@ function loadVue() {
 		`
 	})
 
+    Vue.component('styled-text', {
+        props: ['layer', 'data', 'style'],
+        template: `
+			<div class="instant" v-html="data" :style="style"></div>
+        `
+    })
+
 // data = a function returning the content (actually HTML)
 	Vue.component('raw-html', {
 			props: ['layer', 'data'],
@@ -583,6 +590,14 @@ function loadVue() {
 			v-on:change="player[layer][data] = toValue(document.getElementById('input-' + layer + '-' + data).value, player[layer][data], defaultValue)">
 		`
 	})
+
+    Vue.component('strict-text-input', {
+        props: ['layer', 'data', 'defaultValue'],
+        template: `
+            <input class="instant" :id="'input-' + layer + '-' + data" :value="player[layer][data].toString()" v-on:focus="focused(true)" v-on:blur="focused(false)"
+            v-on:change="player[layer][data] = document.getElementById('input-' + layer + '-' + data).value">
+        `
+    })
 
 	// Updates the value in player[layer][data][0]
 	Vue.component('slider', {